remove constant propagation changes
