+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/plans/logical/QueryPlanConstraints.scala
@@ -122,216 +122,4 @@ trait ConstraintHelper {
     case _: NullIntolerant => expr.children.flatMap(scanNullIntolerantAttribute)
     case _ => Seq.empty[Attribute]
   }
-
-  def normalizeAndReduceWithConstraints(expression: Expression): Expression =
-    reduceWithConstraints(normalize(expression))._1
-
-  private def normalize(expression: Expression) = expression transform {
-    case GreaterThan(x, y) => LessThan(y, x)
-    case GreaterThanOrEqual(x, y) => LessThanOrEqual(y, x)
-  }
-
-  /**
-   * Traverse a condition as a tree and simplify expressions with constraints.
-   * - This functions assumes that the plan has been normalized using [[normalize()]]
-   * - On matching [[And]], recursively traverse both children, simplify child expressions with
-   *   propagated constraints from sibling and propagate up union of constraints.
-   * - If a child of [[And]] is [[LessThan]], [[LessThanOrEqual]], [[EqualTo]], [[EqualNullSafe]],
-   *   propagate the constraint.
-   * - On matching [[Or]] or [[Not]], recursively traverse each children, propagate no constraints.
-   * - Otherwise, stop traversal and propagate no constraints.
-   * @param expression expression to be traversed
-   * @return A tuple including:
-   *         1. Expression: optionally changed expression after traversal
-   *         2. Seq[Expression]: propagated constraints
-   */
-  private def reduceWithConstraints(expression: Expression): (Expression, Seq[Expression]) =
-    expression match {
-      case e @ (_: LessThan | _: LessThanOrEqual | _: EqualTo | _: EqualNullSafe)
-        if e.deterministic => (e, Seq(e))
-      case a @ And(left, right) =>
-        val (newLeft, leftConstraints) = reduceWithConstraints(left)
-        val simplifiedRight = reduceWithConstraints(right, leftConstraints)
-        val (simplifiedNewRight, rightConstraints) = reduceWithConstraints(simplifiedRight)
-        val simplifiedNewLeft = reduceWithConstraints(newLeft, rightConstraints)
-        val newAnd = if ((simplifiedNewLeft fastEquals left) &&
-          (simplifiedNewRight fastEquals right)) {
-          a
-        } else {
-          And(simplifiedNewLeft, simplifiedNewRight)
-        }
-        (newAnd, leftConstraints ++ rightConstraints)
-      case o @ Or(left, right) =>
-        // Ignore the EqualityPredicates from children since they are only propagated through And.
-        val (newLeft, _) = reduceWithConstraints(left)
-        val (newRight, _) = reduceWithConstraints(right)
-        val newOr = if ((newLeft fastEquals left) && (newRight fastEquals right)) {
-          o
-        } else {
-          Or(newLeft, newRight)
-        }
-
-        (newOr, Seq.empty)
-      case n @ Not(child) =>
-        // Ignore the EqualityPredicates from children since they are only propagated through And.
-        val (newChild, _) = reduceWithConstraints(child)
-        val newNot = if (newChild fastEquals child) {
-          n
-        } else {
-          Not(newChild)
-        }
-        (newNot, Seq.empty)
-      case _ => (expression, Seq.empty)
-    }
-
-  private def reduceWithConstraints(expression: Expression, constraints: Seq[Expression]) =
-    constraints.foldLeft(expression)((e, constraint) => reduceWithConstraint(e, constraint))
-
-  private def planEqual(x: Expression, y: Expression) =
-    !x.foldable && !y.foldable && x.canonicalized == y.canonicalized
-
-  private def valueEqual(x: Expression, y: Expression) =
-    x.foldable && y.foldable && EqualTo(x, y).eval(EmptyRow).asInstanceOf[Boolean]
-
-  private def valueLessThan(x: Expression, y: Expression) =
-    x.foldable && y.foldable && LessThan(x, y).eval(EmptyRow).asInstanceOf[Boolean]
-
-  private def valueLessThanOrEqual(x: Expression, y: Expression) =
-    x.foldable && y.foldable && LessThanOrEqual(x, y).eval(EmptyRow).asInstanceOf[Boolean]
-
-  private def reduceWithConstraint(expression: Expression, constraint: Expression): Expression =
-    constraint match {
-      case a LessThan b => expression transformUp {
-        case c LessThan d if planEqual(b, d) && (planEqual(a, c) || valueLessThanOrEqual(c, a)) =>
-          Literal.TrueLiteral
-        case c LessThan d if planEqual(b, c) && (planEqual(a, d) || valueLessThanOrEqual(d, a)) =>
-          Literal.FalseLiteral
-        case c LessThan d if planEqual(a, c) && (planEqual(b, d) || valueLessThanOrEqual(b, d)) =>
-          Literal.TrueLiteral
-        case c LessThan d if planEqual(a, d) && (planEqual(b, c) || valueLessThanOrEqual(b, c)) =>
-          Literal.FalseLiteral
-
-        case c LessThanOrEqual d
-          if planEqual(b, d) && (planEqual(a, c) || valueLessThanOrEqual(c, a)) =>
-          Literal.TrueLiteral
-        case c LessThanOrEqual d
-          if planEqual(b, c) && (planEqual(a, d) || valueLessThanOrEqual(d, a)) =>
-          Literal.FalseLiteral
-        case c LessThanOrEqual d
-          if planEqual(a, c) && (planEqual(b, d) || valueLessThanOrEqual(b, d)) =>
-          Literal.TrueLiteral
-        case c LessThanOrEqual d
-          if planEqual(a, d) && (planEqual(b, c) || valueLessThanOrEqual(b, c)) =>
-          Literal.FalseLiteral
-
-       case c EqualTo d if planEqual(b, d) && (planEqual(a, c) || valueLessThanOrEqual(c, a)) =>
-          Literal.FalseLiteral
-        case c EqualTo d if planEqual(b, c) && (planEqual(a, d) || valueLessThanOrEqual(d, a)) =>
-          Literal.FalseLiteral
-        case c EqualTo d if planEqual(a, c) && (planEqual(b, d) || valueLessThanOrEqual(b, d)) =>
-          Literal.FalseLiteral
-        case c EqualTo d if planEqual(a, d) && (planEqual(b, c) || valueLessThanOrEqual(b, c)) =>
-          Literal.FalseLiteral
-
-        case c EqualNullSafe d
-          if planEqual(b, d) && (planEqual(a, c) || valueLessThanOrEqual(c, a)) =>
-          Literal.FalseLiteral
-        case c EqualNullSafe d
-          if planEqual(b, c) && (planEqual(a, d) || valueLessThanOrEqual(d, a)) =>
-          Literal.FalseLiteral
-        case c EqualNullSafe d
-          if planEqual(a, c) && (planEqual(b, d) || valueLessThanOrEqual(b, d)) =>
-          Literal.FalseLiteral
-        case c EqualNullSafe d
-          if planEqual(a, d) && (planEqual(b, c) || valueLessThanOrEqual(b, c)) =>
-          Literal.FalseLiteral
-
-        case c EqualNullSafe d if planEqual(b, d) => EqualTo(c, d)
-        case c EqualNullSafe d if planEqual(b, c) => EqualTo(c, d)
-        case c EqualNullSafe d if planEqual(a, c) => EqualTo(c, d)
-        case c EqualNullSafe d if planEqual(a, d) => EqualTo(c, d)
-      }
-      case a LessThanOrEqual b => expression transformUp {
-        case c LessThan d if planEqual(b, d) && valueLessThan(c, a) =>
-          Literal.TrueLiteral
-        case c LessThan d if planEqual(b, c) && (planEqual(a, d) || valueLessThanOrEqual(d, a)) =>
-          Literal.FalseLiteral
-        case c LessThan d if planEqual(a, c) && valueLessThan(b, d) =>
-          Literal.TrueLiteral
-        case c LessThan d if planEqual(a, d) && (planEqual(b, c) || valueLessThanOrEqual(b, c)) =>
-          Literal.FalseLiteral
-
-        case c LessThanOrEqual d
-          if planEqual(b, d) && (planEqual(a, c) || valueLessThanOrEqual(c, a)) =>
-          Literal.TrueLiteral
-        case c LessThanOrEqual d if planEqual(b, c) && valueLessThan(d, a) =>
-          Literal.FalseLiteral
-        case c LessThanOrEqual d if planEqual(b, c) && (planEqual(a, d) || valueEqual(a, d)) =>
-          EqualTo(c, d)
-        case c LessThanOrEqual d
-          if planEqual(a, c) && (planEqual(b, d) || valueLessThanOrEqual(b, d)) =>
-          Literal.TrueLiteral
-        case c LessThanOrEqual d if planEqual(a, d) && valueLessThan(b, c) =>
-          Literal.FalseLiteral
-        case c LessThanOrEqual d if planEqual(a, d) && (planEqual(b, c) || valueEqual(b, c)) =>
-          EqualTo(c, d)
-
-        case c EqualTo d if planEqual(b, d) && valueLessThan(c, a) => Literal.FalseLiteral
-        case c EqualTo d if planEqual(b, c) && valueLessThan(d, a) => Literal.FalseLiteral
-        case c EqualTo d if planEqual(a, c) && valueLessThan(b, d) => Literal.FalseLiteral
-        case c EqualTo d if planEqual(a, d) && valueLessThan(b, c) => Literal.FalseLiteral
-
-        case c EqualNullSafe d if planEqual(b, d) && valueLessThan(c, a) => Literal.FalseLiteral
-        case c EqualNullSafe d if planEqual(b, c) && valueLessThan(d, a) => Literal.FalseLiteral
-        case c EqualNullSafe d if planEqual(a, c) && valueLessThan(b, d) => Literal.FalseLiteral
-        case c EqualNullSafe d if planEqual(a, d) && valueLessThan(b, c) => Literal.FalseLiteral
-
-        case c EqualNullSafe d if planEqual(b, d) => EqualTo(c, d)
-        case c EqualNullSafe d if planEqual(b, c) => EqualTo(c, d)
-        case c EqualNullSafe d if planEqual(a, c) => EqualTo(c, d)
-        case c EqualNullSafe d if planEqual(a, d) => EqualTo(c, d)
-      }
-      case a EqualTo b => expression transformUp {
-        case c LessThan d if planEqual(b, d) && planEqual(a, c) => Literal.FalseLiteral
-        case c LessThan d if planEqual(b, c) && planEqual(a, d) => Literal.FalseLiteral
-        case c LessThan d if planEqual(a, d) && planEqual(b, c) => Literal.FalseLiteral
-        case c LessThan d if planEqual(a, c) && planEqual(b, d) => Literal.FalseLiteral
-
-        case c LessThanOrEqual d if planEqual(b, d) && planEqual(a, c) => Literal.TrueLiteral
-        case c LessThanOrEqual d if planEqual(b, c) && planEqual(a, d) => Literal.TrueLiteral
-        case c LessThanOrEqual d if planEqual(a, d) && planEqual(b, c) => Literal.TrueLiteral
-        case c LessThanOrEqual d if planEqual(a, c) && planEqual(b, d) => Literal.TrueLiteral
-
-        case c EqualTo d if planEqual(b, d) && planEqual(a, c) => Literal.TrueLiteral
-        case c EqualTo d if planEqual(b, c) && planEqual(a, d) => Literal.TrueLiteral
-        case c EqualTo d if planEqual(a, d) && planEqual(b, c) => Literal.TrueLiteral
-        case c EqualTo d if planEqual(a, c) && planEqual(b, d) => Literal.TrueLiteral
-
-        case c EqualNullSafe d if planEqual(b, d) =>
-          if (planEqual(a, c)) Literal.TrueLiteral else EqualTo(c, d)
-        case c EqualNullSafe d if planEqual(b, c) =>
-          if (planEqual(a, d)) Literal.TrueLiteral else EqualTo(c, d)
-        case c EqualNullSafe d if planEqual(a, d) =>
-          if (planEqual(b, c)) Literal.TrueLiteral else EqualTo(c, d)
-        case c EqualNullSafe d if planEqual(a, c) =>
-          if (planEqual(b, d)) Literal.TrueLiteral else EqualTo(c, d)
-      }
-      case a EqualNullSafe b => expression transformUp {
-        case c LessThan d if planEqual(b, d) && planEqual(a, c) => Literal.FalseLiteral
-        case c LessThan d if planEqual(b, c) && planEqual(d, a) => Literal.FalseLiteral
-        case c LessThan d if planEqual(a, d) && planEqual(b, c) => Literal.FalseLiteral
-        case c LessThan d if planEqual(a, c) && planEqual(d, b) => Literal.FalseLiteral
-
-        case c LessThanOrEqual d if planEqual(b, d) && planEqual(a, c) => EqualTo(c, d)
-        case c LessThanOrEqual d if planEqual(b, c) && planEqual(a, d) => EqualTo(c, d)
-        case c LessThanOrEqual d if planEqual(a, d) && planEqual(b, c) => EqualTo(c, d)
-        case c LessThanOrEqual d if planEqual(a, c) && planEqual(b, d) => EqualTo(c, d)
-
-        case c EqualNullSafe d if planEqual(b, d) && planEqual(a, c) => Literal.TrueLiteral
-        case c EqualNullSafe d if planEqual(b, c) && planEqual(a, d) => Literal.TrueLiteral
-        case c EqualNullSafe d if planEqual(a, d) && planEqual(b, c) => Literal.TrueLiteral
-        case c EqualNullSafe d if planEqual(a, c) && planEqual(b, d) => Literal.TrueLiteral
-      }
-      case _ => expression
-    }
 }
